# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clusters Metadata API",
    "description": "Canonical API for managed-cluster metadata.",
    "version": "development"
  },
  "paths": {
...
    "/v3/releases": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "List Releases",
        "operationId": "list_releases_v3_releases_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Provider"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReleaseOrderBy",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Create Release",
        "operationId": "create_release_v3_releases_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v3/releases/{release_id}": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Get Release",
        "operationId": "get_release_v3_releases__release_id__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "releases"
        ],
        "summary": "Delete Release",
        "operationId": "delete_release_v3_releases__release_id__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FeatureConstraint": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "operator": {
            "$ref": "#/components/schemas/FeatureOperator"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "operator"
        ],
        "title": "FeatureConstraint"
      },
      "FeatureCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feature_type": {
            "$ref": "#/components/schemas/FeatureType",
            "default": "optional"
          },
          "dependencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FeatureConstraint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          },
          "namespaced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespaced"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FeatureCreate",
        "description": "Schema for creating a feature (used inside release creation)."
      },
...
      "FeatureOperator": {
        "type": "string",
        "enum": [
          "equals",
          "not_equals",
          "in",
          "not_in"
        ],
        "title": "FeatureOperator"
      },
      "FeatureRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feature_type": {
            "$ref": "#/components/schemas/FeatureType",
            "default": "optional"
          },
          "dependencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FeatureConstraint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          },
          "namespaced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespaced"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "FeatureRead",
        "description": "Feature as returned from the API."
      },
...
      "FeatureType": {
        "type": "string",
        "enum": [
          "core",
          "optional"
        ],
        "title": "FeatureType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Provider": {
        "type": "string",
        "enum": [
          "aws",
          "azure"
        ],
        "title": "Provider"
      },
...
      "ReleaseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "reserved_namespaces": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reserved Namespaces"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureCreate"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "provider"
        ],
        "title": "ReleaseCreate"
      },
      "ReleaseDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ReleaseDeleteResponse"
      },
      "ReleaseListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReleaseRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "count",
          "items"
        ],
        "title": "ReleaseListResponse"
      },
      "ReleaseOrderBy": {
        "type": "string",
        "enum": [
          "id",
          "name"
        ],
        "title": "ReleaseOrderBy"
      },
      "ReleaseRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "reserved_namespaces": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reserved Namespaces"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureRead"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "provider"
        ],
        "title": "ReleaseRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}